backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
mysql-connector-python==8.1.0
Werkzeug==2.3.7
flask-bcrypt==1.0.1
redis==5.0.1
gunicorn==21.2.0
